requirements.txt
Pillow~=9.1.1
pystray~=0.19.3
requests~=2.27.1
schedule~=1.1.0
